requirements.txt
aiofiles==0.5.0
aiomysql==0.0.21
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.16.1
aniso8601==7.0.0
anyio==3.3.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.7
dadata==20.7.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.0
fastapi-admin==1.0.3
fastapi-crudrouter==0.8.1
fastapi-jwt-auth==0.5.0
fastapi-mail==0.4.1
fastapi-pagination
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
hiredis==2.0.0
httpcore==0.13.6
httptools==0.1.2
httpx==0.19.0
idna==3.2
iniconfig==1.1.1
iso8601==0.1.16
itsdangerous==1.1.0
Jinja2==3.0.1
MarkupSafe==2.0.1
orjson==3.6.3
packaging==21.0
passlib==1.7.4
password==0.2
pendulum==2.1.2
phonenumbers==8.12.30
pluggy==0.13.1
promise==2.3
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytest==6.2.4
pytest-asyncio==0.15.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.6
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
uvicorn==0.13.4
uvloop==0.16.0
watchgod==0.7
websockets==8.1
